ft_printf

+++ includes/ft_printf.h
@@ -6,14 +6,13 @@
 /*   By: yaajagro <yaajagro@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/12 21:07:36 by yaajagro          #+#    #+#             */
-/*   Updated: 2024/11/17 16:45:46 by yaajagro         ###   ########.fr       */
+/*   Updated: 2024/11/19 21:15:41 by yaajagro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef FT_PRINTF_H
 # define FT_PRINTF_H
 # include <unistd.h>
-# include <stdlib.h>
 # include <stdarg.h>
 
 int	ft_printf(const char *str, ...);
@@ -23,6 +22,6 @@ int	ft_print_nbr(int nbr);
 int	ft_print_hex(unsigned int nbr);
 int	ft_print_hexa(unsigned int nbr);
 int	ft_put_unbr(unsigned int nbr);
-int	ft_print_addres(unsigned long nbr);
+int	ft_print_addres(unsigned long long nbr);
 
 #endif
